fix(ci): satisfy integration quality gates
