+++ rust/tests/suite/context_cortex_phase1.rs
@@ -136,6 +136,7 @@ fn bm25_ingest_content_chunks_increases_doc_count() {
         doc_count: 0,
         doc_freqs: std::collections::HashMap::new(),
         files: std::collections::HashMap::new(),
+        dirs: std::collections::HashMap::new(),
         content_truncated: false,
     };
 
@@ -178,6 +179,7 @@ fn bm25_search_finds_ingested_provider_chunks() {
         doc_count: 0,
         doc_freqs: std::collections::HashMap::new(),
         files: std::collections::HashMap::new(),
+        dirs: std::collections::HashMap::new(),
         content_truncated: false,
     };
 
@@ -218,6 +220,7 @@ fn bm25_mixed_code_and_provider_chunks() {
         doc_count: 0,
         doc_freqs: std::collections::HashMap::new(),
         files: std::collections::HashMap::new(),
+        dirs: std::collections::HashMap::new(),
         content_truncated: false,
     };
 
@@ -265,6 +268,7 @@ fn bm25_external_chunk_count_accurate() {
         doc_count: 0,
         doc_freqs: std::collections::HashMap::new(),
         files: std::collections::HashMap::new(),
+        dirs: std::collections::HashMap::new(),
         content_truncated: false,
     };
 
@@ -314,6 +318,7 @@ fn bm25_ingest_zero_chunks_is_noop() {
         doc_count: 0,
         doc_freqs: std::collections::HashMap::new(),
         files: std::collections::HashMap::new(),
+        dirs: std::collections::HashMap::new(),
         content_truncated: false,
     };
 
@@ -806,6 +811,7 @@ fn end_to_end_provider_to_bm25_search() {
         doc_count: 0,
         doc_freqs: std::collections::HashMap::new(),
         files: std::collections::HashMap::new(),
+        dirs: std::collections::HashMap::new(),
         content_truncated: false,
     };
 
@@ -835,6 +841,7 @@ fn end_to_end_column_pipeline_to_bm25() {
         doc_count: 0,
         doc_freqs: std::collections::HashMap::new(),
         files: std::collections::HashMap::new(),
+        dirs: std::collections::HashMap::new(),
         content_truncated: false,
     };
