+++ rust/tests/suite/context_cortex_phase2.rs
@@ -313,6 +313,7 @@ fn end_to_end_consolidation_populates_all_stores() {
         doc_count: 0,
         doc_freqs: std::collections::HashMap::new(),
         files: std::collections::HashMap::new(),
+        dirs: std::collections::HashMap::new(),
         content_truncated: false,
     };
     let mut edges: Vec<IndexEdge> = Vec::new();
@@ -405,6 +406,7 @@ fn bm25_search_finds_consolidated_provider_data() {
         doc_count: 0,
         doc_freqs: std::collections::HashMap::new(),
         files: std::collections::HashMap::new(),
+        dirs: std::collections::HashMap::new(),
         content_truncated: false,
     };
