+++ rust/tests/suite/context_cortex_phase3.rs
@@ -369,6 +369,7 @@ fn end_to_end_saliency_pipeline() {
         doc_count: 0,
         doc_freqs: std::collections::HashMap::new(),
         files: std::collections::HashMap::new(),
+        dirs: std::collections::HashMap::new(),
         content_truncated: false,
     };
     let mut edges: Vec<IndexEdge> = Vec::new();
